Move packages. Move base pipe to architecture package.
