+++ app/src/main/kotlin/dvachmovie/fragment/movie/MovieBaseFragment.kt
@@ -27,7 +27,7 @@ import dvachmovie.architecture.base.PermissionsCallback
 import dvachmovie.architecture.binding.bindPlayer
 import dvachmovie.architecture.listener.OnSwipeTouchListener
 import dvachmovie.databinding.FragmentMovieBinding
-import dvachmovie.pipe.GetCookiePipe
+import dvachmovie.pipe.network.GetCookiePipe
 import dvachmovie.service.DownloadService
 import dvachmovie.storage.local.MovieDBCache
 import dvachmovie.usecase.MarkCurrentMovieAsPlayedUseCase
