+++ presenter/src/main/kotlin/dvachmovie/architecture/Pipe.kt
@@ -1,0 +1,3 @@
+package dvachmovie.architecture
+
+interface Pipe
