+++ presenter/src/main/kotlin/dvachmovie/architecture/PipeAsync.kt
@@ -1,0 +1,7 @@
+package dvachmovie.architecture
+
+abstract class PipeAsync<in Input> : Pipe {
+
+    abstract suspend fun execute(input: Input)
+
+}
