+++ presenter/src/main/kotlin/dvachmovie/architecture/PipeSync.kt
@@ -1,0 +1,7 @@
+package dvachmovie.architecture
+
+abstract class PipeSync<in Input, out Output> : Pipe {
+
+    abstract fun execute(input: Input): Output
+
+}
