--- presenter/src/main/kotlin/dvachmovie/pipe/Pipe.kt
@@ -1,3 +1,0 @@
-package dvachmovie.pipe
-
-interface Pipe
