--- presenter/src/main/kotlin/dvachmovie/pipe/PipeAsync.kt
@@ -1,7 +1,0 @@
-package dvachmovie.pipe
-
-abstract class PipeAsync<in Input> : Pipe{
-
-    abstract suspend fun execute(input: Input)
-
-}
