--- presenter/src/main/kotlin/dvachmovie/pipe/PipeSync.kt
@@ -1,7 +1,0 @@
-package dvachmovie.pipe
-
-abstract class PipeSync<in Input, out Output> : Pipe{
-
-    abstract fun execute(input: Input): Output
-
-}
