+++ presenter/src/main/kotlin/dvachmovie/pipe/settingsStorage/PutCookiePipe.kt
@@ -1,6 +1,6 @@
 package dvachmovie.pipe.settingsStorage
 
-import dvachmovie.pipe.PipeAsync
+import dvachmovie.architecture.PipeAsync
 import dvachmovie.usecase.settingsStorage.PutCookieUseCase
 import javax.inject.Inject
